chore: rename ipx to local
